pyproject.toml
[tool.setuptools.packages.find]
include = ["core*", "data*", "models*", "serving*"]

[project]
name = "attend-mlops"
version = "0.1.0"
description = "MLOps pipeline for Attend.AI models"
requires-python = ">=3.10"

dependencies = [
    "zenml==0.94.1",
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy-utils",
    "mlflow>=2.19",
    "dvc[s3]",
    "great-expectations",
    "evidently",
    "torch",
    "torchvision",
    "onnxruntime",
    "opencv-python",
    "numpy<2",
    "pandas",
    "scikit-learn",
    "pyyaml",
    "insightface",
    "deepface",
    "huggingface_hub",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]
